% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cld2.R
\name{detect_language}
\alias{detect_language}
\alias{detect_language_multi}
\title{Compact Language Detector 2}
\usage{
detect_language(text, plain_text = TRUE)

detect_language_multi(text, plain_text = TRUE)
}
\arguments{
\item{text}{a string with text to guess.}

\item{plain_text}{if \code{FALSE} then code skips HTML tags and expands HTML entities}
}
\description{
The function \code{\link[=detect_language]{detect_language()}} function returns the best guess of the language of
\code{text} or \code{NA} if the language could not reliablity be determinted.
The function \code{\link[=detect_language_multi]{detect_language_multi()}} returns the top 3 candidate languages with
a probability score, and also the total number of text bytes that was succesfully
classified.
}
